use IProxyMapper in root converter
