--- DUI3-DX/Sdk/Speckle.Converters.Common/IRootToSpeckleConverter.cs
@@ -1,38 +1,0 @@
-using Speckle.Autofac.DependencyInjection;
-using Speckle.Converters.Common.Objects;
-using Speckle.Core.Models;
-using Speckle.InterfaceGenerator;
-
-namespace Speckle.Converters.Common;
-
-[GenerateAutoInterface]
-public class RootToSpeckleConverter : IRootToSpeckleConverter
-{
-  private readonly IFactory<IToSpeckleTopLevelConverter> _toSpeckle;
-
-  public RootToSpeckleConverter(IFactory<IToSpeckleTopLevelConverter> toSpeckle)
-  {
-    _toSpeckle = toSpeckle;
-  }
-
-  public Base Convert(Type type, object target)
-  {
-    try
-    {
-      var objectConverter = _toSpeckle.ResolveInstance(type.Name); //poc: would be nice to have supertypes resolve
-
-      if (objectConverter == null)
-      {
-        throw new NotSupportedException($"No conversion found for {type.Name}");
-      }
-      var convertedObject = objectConverter.Convert(target);
-
-      return convertedObject;
-    }
-    catch (SpeckleConversionException e)
-    {
-      Console.WriteLine(e);
-      throw; // Just rethrowing for now, Logs may be needed here.
-    }
-  }
-}
